Move retry logic into own module and use in Backend
